==> STEP2_ExploreData.Rmd <==
---
title: "STEP2_ExploreData"
author: "Limeng Xie"
date: "5/8/2019"
output:
  html_document:
    df_print: paged
---


```{r}
options(tinytex.verbose = TRUE)
# get the packages we need to use 
setwd("/users/limengxie/Desktop/INFO8000ProjectMomo/")
library(caret)
library(ggplot2)
require(reshape2)
```

```{r}
options(tinytex.verbose = TRUE)
# read the fixed dataset 
data <- read.csv("L88_57_2016_fixed.csv")
head(data)
ncol(data)
# since columns before DIA_STM are just computing parameters or identifications for samples, which are not useful to build models, I delete them
data_model <- data[,-(1:11)]
str(data_model)
write.csv(data_model,file="data_model.csv")
# start to explore these 12 variables
# transfer to tidydata 
melt_data<- melt(data_model,id.var="environment")
head(melt_data)
p <- ggplot(data=melt_data, aes(x=variable, y=value))+
    geom_boxplot(aes(fill=environment))+
    facet_wrap( ~ variable, scales="free")
p
```
## we can see the scale of variables is quite different, which may need to normalize data

## explore the scatterplot 
```{r}
options(tinytex.verbose = TRUE)
featurePlot(x=data_model[,1:6], y=data_model$environment,plot="pairs",auto.key=list(colums=3))
featurePlot(x=data_model[,7:12], y=data_model$environment,plot="pairs",auto.key=list(colums=3))
```
## explore the density plot 
```{r}
options(tinytex.verbose = TRUE)
featurePlot(x=data_model[,1:6], y=data_model$environment, plot="density", scales = list(x = list(relation="free"),                                                                         y = list(relation="free")),
            adjust = 1.5,
            pch = "|",
            layout = c(2, 3))
featurePlot(x=data_model[,c(7:10,12)], y=data_model$environment, plot="density", scales = list(x = list(relation="free"),                                                                         y = list(relation="free")),
            adjust = 1.5,
            pch = "|",
            layout = c(2, 3))
```
## We can see from boxplot, scatterplot and densityplot, three variables TD_MED, ADVT_COUNT, BASAL_COUNT dont show much difference at ww and ws environment, which means they may not be good features to distinguish ww and ws. So I decide to delete 
TD_MED,ADVT_COUNT,BASAL_COUNT. 
```{r}
options(tinytex.verbose = TRUE)
data_model <- data_model[,-c(4,7,8)]
summary(data_model)
str(data_model)
write.csv(data_model, file = "data_model")
```
## Now we have 9 variables in our data set 
DIA_STM: Stem diameter, Area: projected root area, AVG_DENSITY:Ratio between foreground and background pixels of the extracted root， TD_AVG: Mean root tip diameter, RTP_COUNT: number of root tip paths, HYP_DIA: Hypocotyl Diameter, TAP_DIA:Tap root diameter, Environment: well-water, water-stress, Plantmass: the dryweight of above plant tissue. Here, only environment is categorical variable, others are numerical variable.
